Update pypi configuration Added README.txt and CHANGES.txt Fix configuration path for zope.app.securitypolicy and zope.app.security in ftesting.zcml
TODO later:
Find a way to get rid loading the hole zope trunk as dependency.
This package offers a view and a zmi menu which includes the hole
zope as dependency. This package clearly offers a base core
concept and is not intend as a enhancement for rotterdam by default.